Inline simple_expr transform into expr
