+++ src/transform/mod.rs
@@ -3,16 +3,13 @@ pub(crate) mod class_def;
 pub(crate) mod comprehension;
 pub(crate) mod decorator;
 pub(crate) mod destructure;
+pub(crate) mod expr;
 pub(crate) mod for_loop;
 pub(crate) mod gen;
 pub(crate) mod import;
-pub(crate) mod literal;
 pub(crate) mod match_case;
 pub(crate) mod multi_target;
-pub(crate) mod operator;
 pub(crate) mod raise;
-pub(crate) mod simple_expr;
-pub(crate) mod ternary;
 pub(crate) mod truthy;
 pub(crate) mod try_except;
 pub(crate) mod with;
